+++ src/B_read_from_files.py
@@ -127,6 +127,21 @@ def get_parameters_constant(file, sheet_input_constant):
     parameters_constant = parameters_constant.to_dict(orient="dict")
     parameters_constant_units = parameters_constant["Unit"]
     parameters_constant_values = parameters_constant["Value"]
+
+    if "fuel_co2_emission_factor" not in parameters_constant_values.keys():
+        parameters_constant_values.update({"fuel_co2_emission_factor": 2.68})
+        parameters_constant_units.update({"fuel_co2_emission_factor": "kgCO2/l"})
+        logging.warning("Optional parameter `fuel_co2_emission_factor` was not included in the input parameters."
+                        "The CO2 emissions will be calculated with the default value of 2.68 kgCO2/l diesel")
+
+    if "maingrid_co2_emission_factor" not in parameters_constant_values.keys():
+        parameters_constant_values.update({"maingrid_co2_emission_factor": 0.9})
+        parameters_constant_units.update({"maingrid_co2_emission_factor": "kgCO2/kWh"})
+        logging.warning("Optional parameter `maingrid_co2_emission_factor` was not included in the input parameters."
+                        "The CO2 emissions will be calculated with the default value of 0.9 kgCO2/ kWh "
+                        "consumed from the (fully coal-based) maingrid.")
+
+    #print(parameters_constant_values)
     return parameters_constant_units, parameters_constant_values
 
 def get_parameters_sensitivity(file, sheet_input_sensitivity):